Move and fix things
