+++ packages/app/package.json
@@ -14,9 +14,11 @@
     "@visx/geo": "^1.0.0",
     "@visx/grid": "^1.1.0",
     "@visx/group": "^1.0.0",
+    "@visx/responsive": "^1.3.0",
     "@visx/scale": "^1.1.0",
     "@visx/shape": "^1.1.0",
     "@visx/text": "^1.1.0",
+    "@visx/tooltip": "^1.3.0",
     "d3-scale": "^2",
     "date-fns": "^2.15.0",
     "geojson": "^0.5.0",
@@ -47,6 +49,7 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.0.4",
     "@types/classnames": "^2.2.10",
+    "@types/d3-array": "^2.8.0",
     "@types/d3-scale": "^2.2.0",
     "@types/geojson": "^7946.0.7",
     "@types/jest": "^26.0.15",
