+++ packages/app/src/components-styled/line-chart/trends.tsx
