+++ packages/app/src/pages/gemeente/[code]/ziekenhuis-opnames.tsx
@@ -13,7 +13,7 @@ import { createSelectMunicipalHandler } from '~/components/choropleth/select-han
 import { createMunicipalHospitalAdmissionsTooltip } from '~/components/choropleth/tooltips/municipal/create-municipal-hospital-admissions-tooltip';
 import { FCWithLayout } from '~/domain/layout/layout';
 import { getMunicipalityLayout } from '~/domain/layout/municipality-layout';
-import LineChart from '~/components/lineChart';
+import { LineChart } from '~/components-styled/line-chart';
 import { SEOHead } from '~/components/seoHead';
 import siteText from '~/locale/index';
 import {
